[Design] PickerView가 TabBar 위로 보일 수 있도록 디자인 적용
